Prevent cursor moving away
Emacs must position the cursor somewhere.  Its position usually
corresponds to point except for cases where window point would matter.
Inserting text will move point to its end.  In this case, the inserted
text is displayed as an image that's as large as the inner window
dimensions, so the cursor will be put after it.  Usually Emacs solves
this by putting point into the fringe(!) (unless
`overflow-newline-into-fringe` is customized), but if there's no fringe
displayed, it will fall back to the usual behaviour of putting point
after the image and doing some weird scrolling.  Which explains #4.

The easiest way of fixing this is moving point to the beginning or in
other words, right before the image.  This would usually paint a hollow
cursor over the image (which looks like an outline), but since
`cursor-type` is set to `nil`, there won't be any and everything looks
good.

Closes #4.